refactor: move per-request telemetry out of lib into its sinks
lib.rs held the bodies of both event sinks; lift them to the modules that own
each output:

- log_analytics -> analytics.rs (joins RequestEvent/log_request, which drop to
  private now that only log_analytics builds them).
- CfProperties + LocationEvent + maybe_broadcast_location -> new location.rs.

header_str stays in lib as a shared pub(crate) util. Kept the two sinks as
flat sibling modules rather than a telemetry/ parent: they share the
fire-and-forget pattern but no code, so a directory would group for its own
sake. lib.rs drops ~140 lines, back to entrypoint + orchestration + wiring.
