+++ src/lib.rs
@@ -3,12 +3,13 @@ mod authz;
 mod backend_auth;
 mod config;
 mod handlers;
+mod location;
 mod pagination;
 mod source_api;
 mod sts;
 
 use crate::source_api::{ApiAuth, SourceCoopRegistry};
-use analytics::{extract_path_segments, log_request, RequestEvent};
+use analytics::{extract_path_segments, log_analytics};
 use handlers::{AccountListHandler, IndexHandler};
 use multistore::api::response::ErrorResponse;
 use multistore::proxy::{GatewayResponse, ProxyGateway};
@@ -280,11 +281,11 @@ async fn fetch(req: web_sys::Request, env: Env, ctx: Context) -> Result<web_sys:
     // ── Broadcast location to WebSocket viewers ──────────────────
     if let (&http::Method::GET, Some(acct), Some(prod)) = (&parts.method, account, product) {
         if response.status() < 400 && !parts.path.starts_with("/.") {
-            maybe_broadcast_location(
+            location::maybe_broadcast_location(
                 &ctx,
                 &env,
-                LocationEvent {
-                    cf: CfProperties::from_request(&req),
+                location::LocationEvent {
+                    cf: location::CfProperties::from_request(&req),
                     country: header_str(&parts.headers, "cf-ipcountry").to_string(),
                     account: acct.to_string(),
                     product: prod.to_string(),
@@ -324,147 +325,11 @@ fn extract_request_id(headers: &http::HeaderMap) -> String {
         .to_string()
 }
 
-fn header_str<'a>(headers: &'a http::HeaderMap, name: &str) -> &'a str {
+pub(crate) fn header_str<'a>(headers: &'a http::HeaderMap, name: &str) -> &'a str {
     headers
         .get(name)
         .and_then(|v| v.to_str().ok())
         .unwrap_or("")
-}
-
-// ── Analytics ──────────────────────────────────────────────────────
-
-fn log_analytics(
-    env: &Env,
-    headers: &http::HeaderMap,
-    response: &web_sys::Response,
-    method: &http::Method,
-    account: Option<&str>,
-    product: Option<&str>,
-    key: Option<&str>,
-) {
-    let content_type = response
-        .headers()
-        .get("content-type")
-        .ok()
-        .flatten()
-        .unwrap_or_default();
-    let bytes_sent: f64 = response
-        .headers()
-        .get("content-length")
-        .ok()
-        .flatten()
-        .and_then(|v| v.parse().ok())
-        .unwrap_or(0.0);
-
-    log_request(
-        env,
-        &RequestEvent {
-            account_id: account.unwrap_or(""),
-            product_id: product.unwrap_or(""),
-            file_path: key.unwrap_or(""),
-            method: method.as_str(),
-            user_id: header_str(headers, "x-source-user-id"),
-            country: header_str(headers, "cf-ipcountry"),
-            content_type: &content_type,
-            bytes_sent,
-            status_code: response.status() as f64,
-        },
-    );
-}
-
-// ── Location broadcasting ──────────────────────────────────────────
-
-/// Properties extracted from the Cloudflare `request.cf` object.
-#[derive(Default)]
-struct CfProperties {
-    latitude: String,
-    longitude: String,
-    city: String,
-    colo: String,
-}
-
-impl CfProperties {
-    fn from_request(req: &web_sys::Request) -> Self {
-        let cf =
-            js_sys::Reflect::get(req, &wasm_bindgen::JsValue::from_str("cf")).unwrap_or_default();
-        if cf.is_undefined() || cf.is_null() {
-            return Self::default();
-        }
-        let get = |key: &str| -> String {
-            js_sys::Reflect::get(&cf, &wasm_bindgen::JsValue::from_str(key))
-                .ok()
-                .map(|v| {
-                    v.as_string()
-                        .unwrap_or_else(|| v.as_f64().map(|n| n.to_string()).unwrap_or_default())
-                })
-                .unwrap_or_default()
-        };
-        Self {
-            latitude: get("latitude"),
-            longitude: get("longitude"),
-            city: get("city"),
-            colo: get("colo"),
-        }
-    }
-}
-
-struct LocationEvent {
-    cf: CfProperties,
-    country: String,
-    account: String,
-    product: String,
-    key: String,
-    api_base_url: String,
-    api_auth: ApiAuth,
-}
-
-/// Broadcast the request's geolocation to WebSocket viewers via the public-log-stream service.
-/// Runs entirely inside `wait_until` so it never blocks the response.
-fn maybe_broadcast_location(ctx: &Context, env: &Env, event: LocationEvent) {
-    let (Ok(lat), Ok(lon)) = (
-        event.cf.latitude.parse::<f64>(),
-        event.cf.longitude.parse::<f64>(),
-    ) else {
-        return;
-    };
-
-    let Ok(location_ws) = env.service("PUBLIC_LOG_STREAM") else {
-        return;
-    };
-
-    ctx.wait_until(async move {
-        let is_public = source_api::cache::get_or_fetch_product(
-            &event.api_base_url,
-            &event.account,
-            &event.product,
-            &event.api_auth,
-            "",
-            None,
-        )
-        .await
-        .map(|p| p.is_public())
-        .unwrap_or(false);
-        if !is_public {
-            return;
-        }
-
-        let body = serde_json::json!({
-            "lat": lat,
-            "lon": lon,
-            "city": event.cf.city,
-            "country": event.country,
-            "colo": event.cf.colo,
-            "account_id": event.account,
-            "product_id": event.product,
-            "path": event.key,
-        });
-        let mut init = worker::RequestInit::new();
-        init.with_method(worker::Method::Post);
-        init.with_body(Some(wasm_bindgen::JsValue::from_str(&body.to_string())));
-        let _ = location_ws
-            .fetch("https://public-log-stream/location", Some(init))
-            .await;
-    });
 }
 
 // ── CORS ────────────────────────────────────────────────────────────
